Add files via upload
Got mount working, except for a couple variables described in the comments. Image file is one I created on my own machine with no files inside of it. We can put those in later when we get to that part of the project. The imagefile wang has on the project page is garbage.